+++ lib/atoms/adaptive/adaptive_module.dart
@@ -1,3 +1,4 @@
 export 'adaptative_dispatcher.dart';
 export 'adaptive_mdd.dart';
 export 'adaptive_mdd_navigator.dart';
+export 'dashboard_desktop.dart';