+++ lib/molecules/navigation_menu/collapse_state.dart
@@ -1,3 +1,4 @@
+@Deprecated('Ce composant sera bientôt supprimé pour une refonte complète.')
 enum NavigationMenuCollapseState {
   none(0.0),
   collapsed(92.0),