+++ lib/molecules/navigation_menu/destination_data.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+@Deprecated('Ce composant sera bientôt supprimé pour une refonte complète.')
 class DestinationData {
   final IconData? icon;
   final String label;